services/text-interpreter/requirements.txt
grpcio 
grpcio-tools 
fastapi